Move all exceptions into errors.rb
